requirements.txt
azure-functions
azure-cosmos
azure-storage-blob
pandas
python-dateutil
azure-storage-queue

